requirements.txt
customtkinter==5.2.1
sympy==1.12
numpy==1.24.3
matplotlib==3.7.1
reportlab==4.0.4
pillow==10.0.0
scipy==1.11.1